requirements.txt
torch
torchvision
pytesseract
opencv-python-headless
streamlit
Pillow
numpy